+++ source/core/parser.c
@@ -313,7 +313,7 @@ static bool is_numeric_type(ValueType type) {
 static bool is_comparable_type(ValueType type) {
     return type == TYPE_NUMBER || type == TYPE_STRING || 
            type == TYPE_BOOLEAN || type == TYPE_NONE || 
-           type == TYPE_FUNCTION;
+           type == TYPE_FUNCTION || type == TYPE_TABLE;
 }
 
 // returns the string representation of a binary operator token
@@ -882,6 +882,9 @@ static ValueType infer_binary_type(Parser* parser, ASTNode* node) {
                 return TYPE_BOOLEAN;
             }
             if (left_type == TYPE_FUNCTION && right_type == TYPE_FUNCTION) {
+                return TYPE_BOOLEAN;
+            }
+            if (left_type == TYPE_TABLE && right_type == TYPE_TABLE) {
                 return TYPE_BOOLEAN;
             }
             if (!is_comparable_type(left_type) || !is_comparable_type(right_type)) {